+++ python/IECoreRenderManTest/RendererTest.py
@@ -2180,6 +2180,69 @@ class RendererTest( GafferTest.TestCase ) :
 			"identifier:id", [ 1 ]
 		)
 
+	def testGroupingMembership( self ) :
+
+		with IECoreRenderManTest.RileyCapture() as capture :
+
+			renderer = GafferScene.Private.IECoreScenePreview.Renderer.create(
+				self.renderer,
+				GafferScene.Private.IECoreScenePreview.Renderer.RenderType.Batch
+			)
+
+			renderer.object(
+				"/sphere", IECoreScene.SpherePrimitive( 1 ),
+				renderer.attributes( IECore.CompoundObject( {
+					"ri:grouping:membership" : IECore.StringData( "groupA groupB" )
+				} ) )
+			)
+
+			del renderer
+
+		# Memberships specified by the user should have been combined with those
+		# generated automatically for light linking.
+		self.__assertParameterEqual(
+			next( x for x in capture.json if x["method"] == "CreateGeometryInstance" )["attributes"]["params"],
+			"grouping:membership", [ "defaultShadowGroup groupA groupB" ]
+		)
+
+	def testGroupingMembershipMergedWithShadowLinking( self ) :
+
+		with IECoreRenderManTest.RileyCapture() as capture :
+
+			renderer = GafferScene.Private.IECoreScenePreview.Renderer.create(
+				self.renderer,
+				GafferScene.Private.IECoreScenePreview.Renderer.RenderType.Batch
+			)
+
+			lightAttributes = renderer.attributes(
+				IECore.CompoundObject( {
+					"ri:light" : IECoreScene.ShaderNetwork(
+						shaders = { "output" : IECoreScene.Shader( "PxrDomeLight", "ri:light" ) },
+						output = "output",
+					),
+				} )
+			)
+
+			light = renderer.light( "light", None, lightAttributes )
+
+			object = renderer.object(
+				"/sphere", IECoreScene.SpherePrimitive( 1 ),
+				renderer.attributes( IECore.CompoundObject( {
+					"ri:grouping:membership" : IECore.StringData( "groupA groupB" )
+				} ) )
+			)
+
+			object.link( "shadowedLights", { light } )
+
+			del renderer
+
+		# Memberships specified by the user should have been combined with those
+		# generated automatically for shadow linking.
+		self.__assertParameterEqual(
+			next( x for x in capture.json if x["method"] == "ModifyGeometryInstance" )["attributes"]["params"],
+			"grouping:membership", [ "shadowGroup0 groupA groupB" ]
+		)
+
 	def testShaderSubstitutions( self ) :
 
 		with IECoreRenderManTest.RileyCapture() as capture :
@@ -2323,6 +2386,42 @@ class RendererTest( GafferTest.TestCase ) :
 							cam['projection']['params']['params'], "focalLength", [focalLength  * flWorldScale],
 							tolerance = 1e-7
 						)
+
+	def testCustomCameraParameters( self ) :
+
+		camera = IECoreScene.Camera()
+		camera.setProjection( "perspective" )
+		camera.parameters()["ri:radial1"] = 1.0
+		camera.parameters()["ri:apertureAngle"] = 45.0
+		camera.parameters()["ri:apertureDensity"] = 5.0
+		camera.parameters()["ri:apertureNSides"] = 3
+		camera.parameters()["ri:apertureRoundness"] = 0.0
+		camera.parameters()["ri:dofaspect"] = 1.0
+		camera.parameters()["ai:ignoreMe"] = 1.0
+
+		with IECoreRenderManTest.RileyCapture() as capture :
+
+			renderer = GafferScene.Private.IECoreScenePreview.Renderer.create(
+				"RenderMan",
+				GafferScene.Private.IECoreScenePreview.Renderer.RenderType.Batch
+			)
+
+			renderer.camera(
+				"camera", camera, renderer.attributes( IECore.CompoundObject() )
+			)
+
+			del renderer
+
+		camera = next( x for x in capture.json if x["method"] == "CreateCamera" )
+		self.__assertParameterEqual( camera["projection"]["params"]["params"], "radial1", [ 1.0 ] )
+		self.__assertParameterEqual( camera["properties"]["params"], "apertureAngle", [ 45.0 ] )
+		self.__assertParameterEqual( camera["properties"]["params"], "apertureDensity", [ 5.0 ] )
+		self.__assertParameterEqual( camera["properties"]["params"], "apertureNSides", [ 3 ] )
+		self.__assertParameterEqual( camera["properties"]["params"], "apertureRoundness", [ 0.0 ] )
+		self.__assertParameterEqual( camera["properties"]["params"], "dofaspect", [ 1.0 ] )
+		for parameterList in ( camera["properties"]["params"], camera["projection"]["params"]["params"] ) :
+			self.__assertNotInParameters( parameterList, "ai:ignoreMe" )
+			self.__assertNotInParameters( parameterList, "ignoreMe" )
 
 	def runOverscanTest( self, res, pixelsTop, pixelsBottom, pixelsLeft, pixelsRight ) :